# MESMER-X: compile expression