<!-- Feel free to remove check-list items aren't relevant to your change -->

 - [ ] Closes #xxx
 - [ ] Tests added
 - [ ] Fully documented, including `CHANGELOG.rst`

Another way to speed up MESMER-X. 


```python
import numpy as np

c1 = 2.558745
tas = np.array([1.3, 4])

a = "c1"
a_compiled = compile(a, "", "eval")

b = "c1 * tas"
b_compiled = compile(b, "", "eval")

%timeit eval(a)
%timeit eval(a_compiled)
%timeit eval(b)
%timeit eval(b_compiled)
```


```
5.23 μs ± 70.8 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
411 ns ± 8.39 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
9.53 μs ± 615 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
1.42 μs ± 41.7 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```