@mathause 2025-02-04
## Speedup


**py3.13 ubuntu-latest**

Before (3m 43s total)

```
56.41s call    tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1_2ndfit-True-False]
29.68s call    tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1-False-False]
2.75s call     tests/unit/test_mesmer_x_first_guess.py::test_first_guess_truncnorm
```

After (3m 26s)

```
46.40s call     tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1_2ndfit-True-False]
24.20s call     tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1-False-False]
```

---

**py3.13 windows-latest**


Before (9m 52s)
```

75.44s call     tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1_2ndfit-True-False]
39.01s call     tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1-False-False]
3.44s call     tests/unit/test_mesmer_x_first_guess.py::test_first_guess_truncnorm
```

After (7m 37s)

```
60.77s call     tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1_2ndfit-True-False]
30.76s call     tests/integration/test_calibrate_mesmer_x.py::test_calibrate_mesmer_x[ssp585-tasmax-norm(loc=c1 + c2 * __tas__, scale=c3)-expr1-False-False]
```


---

* Before numbers from #611
* this obviously also includes #613
* `test_first_guess_truncnorm` no longer in the 10 slowest tests