@mathause 2025-02-04
@yquilcaille & @l-pierini: this should speed up MESMER-X a bit more. I think that's it for now with the performace-related PRs.